+++ app/src/main/java/dev/esnault/bunpyro/data/repository/settings/ISettingsRepository.kt
@@ -1,5 +1,6 @@
 package dev.esnault.bunpyro.data.repository.settings
 
+import dev.esnault.bunpyro.domain.entities.grammar.AllGrammarFilter
 import dev.esnault.bunpyro.domain.entities.settings.FuriganaSetting
 import dev.esnault.bunpyro.domain.entities.settings.NightModeSetting
 
@@ -9,4 +10,7 @@ interface ISettingsRepository {
 
     suspend fun getFurigana(): FuriganaSetting
     suspend fun setFurigana(setting: FuriganaSetting)
+
+    suspend fun setAllGrammarFilter(filter: AllGrammarFilter)
+    suspend fun getAllGrammarFilter(): AllGrammarFilter
 }