+++ app/src/main/java/dev/esnault/bunpyro/data/repository/settings/SettingsRepository.kt
@@ -3,6 +3,9 @@ package dev.esnault.bunpyro.data.repository.settings
 import android.content.Context
 import androidx.core.content.edit
 import androidx.preference.PreferenceManager
+import dev.esnault.bunpyro.data.mapper.settings.AllGrammarFilterFromStringMapper
+import dev.esnault.bunpyro.data.mapper.settings.AllGrammarFilterToStringMapper
+import dev.esnault.bunpyro.domain.entities.grammar.AllGrammarFilter
 import dev.esnault.bunpyro.domain.entities.settings.FuriganaSetting
 import dev.esnault.bunpyro.domain.entities.settings.NightModeSetting
 
@@ -24,6 +27,18 @@ class SettingsRepository(context: Context) : ISettingsRepository {
     override suspend fun setFurigana(setting: FuriganaSetting) {
         sharedPreferences.edit {
             putString("furigana_default", setting.value)
+        }
+    }
+
+    override suspend fun getAllGrammarFilter(): AllGrammarFilter {
+        val value = sharedPreferences.getString("all_grammar_filter", null)
+        return AllGrammarFilterFromStringMapper().map(value)
+    }
+
+    override suspend fun setAllGrammarFilter(filter: AllGrammarFilter) {
+        val value = AllGrammarFilterToStringMapper().map(filter)
+        sharedPreferences.edit {
+            putString("all_grammar_filter", value)
         }
     }
 }