+++ app/src/main/java/dev/esnault/bunpyro/domain/entities/JLPT.kt
@@ -1,8 +1,8 @@
 package dev.esnault.bunpyro.domain.entities
 
 
-enum class JLPT {
-    N5, N4, N3, N2, N1;
+enum class JLPT(val level: Int) {
+    N5(5), N4(4), N3(3), N2(2), N1(1);
 
     companion object {
         operator fun get(level: Int): JLPT {